[ApiRsMonitoraggio] Riviste interfacce OpenAPI 3 dei servizi di configurazione e monitoraggio per utilizzare un discriminator interno agli oggetti elencati nelle strutture oneOf,	come richiesto dalla specifica OpenAPI "https://swagger.io/docs/specification/data-models/inheritance-and-polymorphism/". Nella precedente versione il discriminator utilizzato era un claim esterno agli oggetti riferiti nella struttura oneOf.

+++ core/src/org/openspcoop2/core/registry/driver/db/DriverRegistroServiziDB_LIB.java
@@ -4320,12 +4320,12 @@ public class DriverRegistroServiziDB_LIB {
 	}
 
 	
-	public static long CRUDResource(int type, AccordoServizioParteComune as,Resource resource, Connection con, long idAccordo) throws DriverRegistroServiziException {
+	public static int CRUDResource(int type, AccordoServizioParteComune as,Resource resource, Connection con, long idAccordo) throws DriverRegistroServiziException {
 		PreparedStatement updateStmt = null;
 		String updateQuery;
 		PreparedStatement selectStmt = null;
 		ResultSet selectRS = null;
-		long n = 0;
+		int n = 0;
 		if (idAccordo <= 0)
 			new Exception("[DriverRegistroServiziDB_LIB::CRUDResource] ID Accordo non valido.");
